Speichern und Mittelwert Rechnung in Evaluation Hinzugefügt und backend datei geleert
